+++ Demo_ver/Demo_ver/src/main/java/PibuStory/DemoVerApplication.java
@@ -1,0 +1,14 @@
+package PibuStory;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+
+@SpringBootApplication
+public class DemoVerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DemoVerApplication.class, args);
+	}
+
+}
