--- Demo_ver/Demo_ver/src/main/java/PibuStory/Demo_ver/DemoVerApplication.java
@@ -1,13 +1,0 @@
-package PibuStory.Demo_ver;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DemoVerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DemoVerApplication.class, args);
-	}
-
-}
