+++ Demo_ver/Demo_ver/src/main/resources/application.properties
@@ -1,1 +1,8 @@
 spring.application.name=Demo_ver
+
+# db ??
+spring.datasource.url=jdbc:mysql://localhost:3306/mydb
+spring.datasource.username=root
+spring.datasource.password=password
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.database-platform=org.hibernate.dialect.MySQL5Dialect
