--- Demo_ver/Demo_ver/src/main/resources/templates/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-    <h1>얼굴 피부 분석 페이지</h1>
-    <form method="post" enctype="multipart/form-data" th:action="@{/analyze}">
-        <input type="file" name="file" accept="image/*" required />
-        <button type="submit">검사시작</button>
-    </form>
-</body>
-</html>
