--- Demo_ver/Demo_ver/src/main/resources/templates/result.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-  <title>Analysis Result</title>
-</head>
-<body>
-  <h1>Analysis Result</h1>
-  <p th:text="${result}"></p>
-  <a href="/">다시 검사하기</a>
-</body>
-</html>
