+++ Demo_ver/Demo_ver/src/test/java/PibuStory/DemoVerApplicationTests.java
@@ -1,0 +1,13 @@
+package PibuStory;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DemoVerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
