--- Demo_ver/Demo_ver/src/test/java/PibuStory/Demo_ver/DemoVerApplicationTests.java
@@ -1,13 +1,0 @@
-package PibuStory.Demo_ver;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DemoVerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
